removed mothra-c.c bc it is redundant and updated build scripts.  also renamed mothra-c.h to mothra.h
