+++ bindings/Makefile
@@ -2,12 +2,6 @@ SHELL := /bin/sh
 
 include ../config.mk
 
-rust-bindings:
-	cd $(RBIND_DIR) && make $@
-
-c-bindings:
-	cd $(CBIND_DIR) && make $@
-
 dotnet-bindings:
 	cd $(DBIND_DIR) && make $@
