+++ bindings/c/mothra.h
