+++ bindings/java/mothra-jni.h
@@ -1,7 +1,7 @@
 #include <jni.h>
 #ifndef _MOTHRA_JNI_H_
 #define _MOTHRA_JNI_H_
-#include "mothra-c.h"
+#include "mothra.h"
 #ifdef __cplusplus
 extern "C" {
 #endif
