+++ examples/c/Makefile
@@ -8,7 +8,7 @@ INCLUDES:=-I$(CBIND_DIR)
 .PHONY : examples clean
 
 c-examples:
-	$(CC) example.c $(CBIND_DIR)/mothra-c.c $(INCLUDES) -L$(OUT_DIR)/ -lmothra $(OS_CFLAGS) $(CFLAGS) -o $(OUT_DIR)/c-example
+	$(CC) example.c $(INCLUDES) -L$(OUT_DIR)/ -lmothra $(OS_CFLAGS) $(CFLAGS) -o $(OUT_DIR)/c-example
 ifeq ($(OS), darwin)
 	install_name_tool -change $(OUT_DIR)/release/deps/libmothra.dylib @rpath/libmothra.dylib $(OUT_DIR)/c-example
 	install_name_tool -add_rpath @loader_path $(OUT_DIR)/c-example
